Onwards it goes... ;-)
Optimized the mixer layout a bit.
Added labels to most things. Not 100% sure about this....
Pads now have velocity on the Y axis and can be transposed in steps of
16. They send on channel 10 on a separate input from:
the Keys. Added new page with a keyboard sending on channel 1, also
transposable (in octaves).
Two midi inputs now, "TouchOSC Keys" and "TouchOSC Pads".
Keys also have velocity on the Y-axis.
The controller Script is pretty much complete now, but not optimised
yet...